(PC-4682) : simplify get_existing_object
`get_existing_object` had two major issues.
It was redoing the work of the ORM.
It was opening a ghost DB connection.

The ghost DB connection broke the tests because it was checking
the object outside the scope of the current connection.
That should be fine in production as it is unlikely the object was
created earlier during the process.
However, it may lead or have lead to staled DB connection at some
point since the connection does not seem to be closed anywhere.