(PC-4708) Do not use nltk stopwords list when searching
We use nltk stopwords to filter out stopwords before querying the
database for a search. This is mostly useless because PostgreSQL
already does the same filtering with a very similar list of
stopwords. Example:

    # select to_tsvector('french_unaccent', 'Le chocolat est bon')
             @@ to_tsquery('french_unaccent', 'le:*');
    NOTICE:  text-search query contains only stop words or doesn't contain lexemes, ignored
     ?column?
    ----------
     f

There are only two words that PostgreSQL does not consider as
stopwords [1]: "les" and "ils". I would be inclined to just use
PostgreSQL stopwords to simplify our code, but let's be conservative.

Removing nltk will avoid the need to download the stopwords, which we
did on every deployment. The download sometimes failed, which made the
deployment fail too.

[1] As of version 12.3, the list of PostgreSQL stopwords is here:
    https://github.com/postgres/postgres/blob/5060275aa8a1ead56e0a41308d7a43049a6cbe43/src/backend/snowball/stopwords/french.stop