(PC-4708) Simplify tests for filter_offerers_with_keywords_string
We had 4 tests that had the almost same 50 lines of code. Because of
all the boilerplate, finding out the differences between each test was
hard. So was reading the assertions. Also, I was confused by variable
names: since the implementation does not care about the presence or
absence of offers, let's not create any offer at all.