(PC-5630) Revert "Add an input base model that contains the right configuration and use it in offer serialization"
This reverts commit 18264761ec525ca0ca0c1ecca8b15e0806c77515.

We cannot safely apply this commit because some routes may send more
data than pydantic models accept. For example,
`patch_all_offers_active_status` sends a "page" field which does not
appear in the body model. The request is hence rejected by pydantic
with a 400 error code.

This field is useless and the frontend should stop sending it... But
there may be other fields and other routes with the same issue, so
let's play it safe.