(PC-5668) provider_api: Do not silently ignore non-JSON responses of stock providers
I found out that Praxiel returns invalid JSON when a price is less
than 1 €, something like:

    {"ref": "123456789", "available": 31, "price": .95}

It should be `0.95`, not `.95`.

When receiving invalid JSON, we were acting as if the provider
returned an empty list of stocks. We then proceeded to update
`VenueProvider.lastSyncDate` as if everything was fine. That
was wrong!

We now let the possible exception bubble up so that the invalid
response is not ignored: we will NOT update `lastSyncDate` anymore.