(PC-5380) : Add a transaction decorator
This will ensures changes made to the DB will be commited if
everything went fine and rollbacked in case of an exception.
It is aimed to be used with `session.flush` instead of
`session.commit`.

I whish I could set that on teardown_request but I'm unsure of
the possible side effects at the moment.

The decorator should allow code to be moved out of pseudo autocommit
mode of `repository.repository.save()`.