(PC-5168) : Entrypoint to redict url to the native app
This is a proof of concept to work around an issue with
links to the native application that are not correctly handled
by MailJet.
Instead of sending a link to the native application, we build a
link to the backend which will rediect to the native application.
The expected outcome being:
- the user clicks on the mail's link.
- the mobile browser opens and sends the request to the server
- the server redirects to the native application urls
- the mobile browser get the redirect and forwards the request to
the native application.