(PC-6627) providers: Reindex offers within offer/stock synchronization
Previously, when synchronizing offers and stock against a provider, we
enqueued the reindexation of the *whole* venue. This is very long and
too expensive when only a few offers have been created or updated and
thus need to be reindexed.

We now ask for the reindexation of offers that have been created or
updated only.

I am a coward, so the new behaviour can be deactivated through a
feature flag `ENABLE_WHOLE_VENUE_PROVIDER_ALGOLIA_INDEXATION`.