mails: Use our requests wrapper when using the Mailjet HTTP API
HTTP requests to the Mailjet API were not logged. We don't know if
they are slow... That would be a good first step before potentially
making e-mail sending asynchronous.

I chose to define a subclass of the original Mailjet client class. It
is a bit verbose because the original code is not really pluggable.
The alternative would be to monkey-patch the `requests` imported in
`mailjet_rest.client` but I find that even more unpleasant.