(PC-7371) : favorite's offer's same prices should be seen as one
If an offer has several times the same price, the favorite would
show a "start from" tag while there's in fact a unique one.
Instead of counting prices, compute the max and min prices.
If max and min differs, we should display the minimum price as
"start from" while if max and min are the same, just announce it
as a single price.