(PC-7008): LocalClass is a nullable column
Spectree model was not aligned with corresponding SQLAlchemy model:
provider.localClass is a nullable column but was not in spectree

This hasn't been a pb since there was no such provider anymore, they all had localClass filled.
Until the new version of "api fournisseur" providers:
they do not have a personal localClass, only there endpoint's url differ, saved in the new provider.apiUrl column.
Offers synchronized with it were not visible in the webapp, getting a 500. The new providers have not been activated because of that.