(PC-7830) bookings: Optimize `find_soon_to_be_expiring_booking_ordered_by_user()`
The previous version of the query took about 60 seconds on staging. By
applying some filters first (before the JOIN on the "offer"), the
query is much quicker: around 200 ms on staging.

The `Booking.expirationDate` hybrid property may have been a good
idea, but here it made us perform an under-optimized SQL query. To
avoid this in the future, let's remove it. We can keep a "standard"
property for this (and hope that the caller will have joinedload'ed
the offer to avoid N+1 SQL queries).