(PC-7904) admin: Fix inclusion of first/last name in beneficiary creation/edition
Super-admins should be able to see form fields for the first name and
last name when creating or editing a beneficiary (see 48883f3594e603).

It did not work because the form class (and its fields) was cached by
Flask-Admin when the admin view was registered, i.e. when there was no
logged-in user.

Not populating the form cache seemed hard (as other things in
Flask-Admin expected the cache to be there), so here we simply
override a couple of functions to avoid using the cache.