(PC-7924) logging: Introduce a StructureMessage class for our logs
The previous implementation did not work if a logger was defined
before we had called `install_logging()`. The `extra` argument
was always empty in generated JSON logs.

The new approach uses the (standard) log record factory, but it's not
enough because the `Logger.makeRecord()` does not pass extra arguments
to this factory. So we need to introduce a special class to use in our
own logs, like this:

    from pcapi.core.logging import M
    logger.info(M("Frobulated a blob", blob=blob.id, reason=reason))

External logs are still generated as JSON with context information
such as the current user and correlation id (thanks to the formatter).
But they will always have an empty `extra` attribute. That's fine.