(PC-8591) : add favorites/count to the native API
Since the native application displays a badge with the number of
favorites for a user, it should not need to fetch the full favorites
data since they are costly to compute.
Instead of loading the entire favorites data, we add an endpoint
to return the number of favorites which is fast and simple to get.