(PC-8709): fix stale offer state in native API bookings endpoint
The native application uses the wrong url for the bookings.
While this will be fixed in later version, the backend will work
around that issue by returning the booking's url at
`booking.stock.offer.url`.
This alters the offer's url which confuses SQLAlchemy resulting in
the offer to be removed from the session to ensure it will not
be commited be error.