(PC-8909) : fix email validation when user has no idcheck token left
When adding the limitation on IdCheck token quantity for a given
user, the email validation route was forgotten and had sometime
an uncaught exception when trying to create a new token and the
user is not allowed to create one anymore.
Let's catch that exception and return an empty token just like
a non eligible user.