(PC-8910) : Don't let expired tokens be tested as recaptchas
licence_verify was trying to get an active and valid token before
falling back to recaptcha.
However it creates false warning about broken recaptcha.
Therefore the API will now try to sort out whether it's a token
or a recaptcha and then will check the token is valid in a second
step.