(PC-9043) Faster `generate_payments` command
Problems we wanted to solve
===========================

The command had two issues:

  - It handled Payment objects instead of query objects. That caused a
  lot SQL queries to be issued when adding related PaymentStatus or
  updating Payment objects.

  - Some parts did not use joinedload which caused many N+1 queries.

With a growing number of bookings (and thus a growing number of
payments to generate), the script now takes many hours. Also, if the
script stops before its normal end, manual operations are necessary
and it's hard.

Proposed solution
=================

The new version introduces an indexed `Payment.batchDate` column that
indicates that a payment is handled in a particular batch. As such,
all steps of the script use this column: we don't have to keep all
objects in memory, we just query the database when we need to fetch
data for the current batch of payments. And if the script ends
prematurely, it can be run again with the same batch date.

It's almost equivalent to `Payment.paymentMessage`, except that:

  - `paymentMessage` is set too late in the process and may
  actually *not* be set if an error occurs early. `Payment.batchDate`
  is set as soon as a payment is generated.

  - `paymentMessage` is set only once. When the script is run 2 weeks
  later, some old payments (linked to a former payment message) that
  could not be sent are retried and their `paymentMessage` is the
  same. Their `Payment.batchDate`, though, will be updated.

Ditto for `Payment.transactionLabel` that we could not use for the
same reason. I choose not to change the meaning of these two
attributes, not exactly knowing what the consequences would be.

Other changes include:

- better use of joinedload to avoid N+1 queries;
- batch generation of CSV data, when the query would be too big to be
  fully loaded in memory (especially since we need to JOIN half the
  world);
- `generate_new_payments` now iterates only over venues for which we
  may generate a new payment (i.e. venues that have used bookings with
  no payments).
- use of "INSERT INTO ... (SELECT ...)" to efficiently bulk-insert
  data;
- more logs so that we can more precisely analyze the slowest parts.

... because some parts of the script are still slow. Most notably the
`generate_new_payments()` function. Further optimizations are planned.