(PC-9729) : Recompute dnBookedQuantity for each booking batch cancelled
This fixes the fact that only the latest batch of cancelled bookings
was used to recompute dnBookedQuantity on the associated stock.
We also want to commit the recomputed dnBookedQuantity along with
the cancelled bookings so that if something goes south for whatever
reason during the process, the already cancelled bookings will be
aligned with the stock's dnBookedQuantity.