(PC-10624) : improve mediation tests reliability
Mediation tests are creating image files with a 4 char name.
At some point, names are reused which makes the file count unreliable.
In order to secure it, tests now call `clear_tests_assets_bucket`
to clean the asset directory before running the tests.
Cleaning the repository before running the tests allows the files
to remain on disk if the test fails.