(PC-11330) feat:account: update email
Update POST /profile (native v1): add two input parameters: email and
password. They are both needed to start the email update process. The
first one is the new email adress and the password is a security option.

The email is not updated immediately: an email with a validation
link is sent to this new email adress.

A new email address is invalid if an existing email exists.