(PC-11229) hot-fix: Do not check eligibility when creating deposit
Hot fix in prod: if a user turns 19 after having applied to DMS or Jouve, user.eligibility is None when we create the deposit. Temporary force the eligibility in activate_beneficiary function which is called in DMS, Jouve and Admin flows. We will think about a better solution after for pass15-17