(PC-11330) feat:account:email update: save TTL
Save the token's TTL since we are using a JWT that is not saved into the
database. This prevents from having two (or more) active token for the
same user. And we will be able to send this information to the front end
(needed for an incoming feature).

Token TTL is saved using Redis: is not a vital information and redis'
tools like incr/unlink makes things easier.